Prohibit calling builder methods accepting callbacks multiple times
Previously, multiple invocations of Transition.Builder.before, for
example, would override previously associated callbacks, which is almost
certainly an error considering how the FSM instantiated procedure is
supposed to (designed to) happen in one statement. As such, throw an
IllegalStateException if the user tries to call before, after,
beforeStateChange, or afterStateChange more than once, or call onExit or
onEnter with the same state more than once.